requirements.txt
requests>=2.31.0
feedparser>=6.0.10
sentence-transformers>=2.2.2
arxiv>=2.1.0
PyYAML>=6.0.1
beautifulsoup4>=4.12.0
pypdf>=4.0.0
openai>=1.0.0
